Remove references to GenericWords
