+++ __mocks__/firebase/auth.ts
@@ -1,0 +1,6 @@
+export const getAuth = jest.fn(() => ({
+  currentUser: {
+    getIdToken: jest.fn(async () => ''),
+  },
+}));
+
