+++ index.tsx
@@ -14,12 +14,7 @@ import adminRouter from './src/backend/routers/adminRouter';
 import testRouter from './src/backend/routers/testRouter';
 import errorHandler from './src/backend/middleware/errorHandler';
 import afterRes from './src/backend/middleware/afterRes';
-import {
-  onCreateUserAccount,
-  onDeleteUser,
-  onAssignUserToEditingGroup,
-  onUpdatePermissions,
-} from './src/backend/functions/users';
+import { onCreateUserAccount, onDeleteUser, onUpdatePermissions } from './src/backend/functions/users';
 import { onRequestDeleteDocument, onUpdateDocument } from './src/backend/functions/documents';
 import { onTwitterAuth, onTwitterCallback, onDeleteConstructedTermPoll } from './src/backend/controllers/polls';
 import { onMediaSignedRequest } from './src/backend/controllers/media';
@@ -60,7 +55,6 @@ server.use(errorHandler);
 
 // Firebase Functions
 export const createUserAccount = onCreateUserAccount;
-export const assignUserToEditingGroup = onAssignUserToEditingGroup;
 export const updatePermissions = onUpdatePermissions;
 export const requestDeleteDocument = onRequestDeleteDocument;
 export const deleteConstructedTermPoll = onDeleteConstructedTermPoll;
