+++ src/backend/controllers/examples.ts
@@ -11,8 +11,8 @@ import { exampleSuggestionSchema } from 'src/backend/models/ExampleSuggestion';
 import { exampleSchema } from 'src/backend/models/Example';
 import { wordSchema } from 'src/backend/models/Word';
 import { deleteAudioPronunciation } from 'src/backend/controllers/utils/MediaAPIs/AudioAPI';
+import { DICTIONARY_APP_URL } from 'src/backend/config';
 import SuggestionTypes from '../shared/constants/SuggestionTypes';
-import { DICTIONARY_APP_URL } from '../config';
 import { packageResponse, handleQueries, updateDocumentMerge } from './utils';
 import { searchExamplesRegexQuery, searchForAssociatedExampleSuggestions } from './utils/queries';
 import { findExampleSuggestionById } from './exampleSuggestions';
