+++ src/backend/middleware/validateWordMerge.ts
@@ -2,7 +2,6 @@ import mongoose from 'mongoose';
 import { Response, NextFunction } from 'express';
 import Joi from 'joi';
 import * as Interfaces from 'src/backend/controllers/utils/interfaces';
-import { findGenericWordById } from '../controllers/genericWords';
 import { findWordSuggestionById } from '../controllers/wordSuggestions';
 
 const { Types } = mongoose;
@@ -21,8 +20,7 @@ export default async (
   next: NextFunction,
 ): Promise<Response<any> | void> => {
   const { body: finalData, user, mongooseConnection } = req;
-  const suggestionDoc: any = ((await findWordSuggestionById(finalData.id, mongooseConnection))
-  || (await findGenericWordById(finalData.id)));
+  const suggestionDoc: any = await findWordSuggestionById(finalData.id, mongooseConnection);
   req.suggestionDoc = suggestionDoc;
 
   if (!user || (user && !user.uid)) {
