+++ src/hooks/useFirebaseConfig.ts
@@ -1,4 +1,27 @@
-import { PROD_FIREBASE_CONFIG, STAGING_FIREBASE_CONFIG } from '../config';
+/**
+ * Establish Firebase app with appropriate configs
+ */
+export const PROD_FIREBASE_CONFIG = {
+  apiKey: 'AIzaSyBdVZrCXPN5wZiNbbVkqz9SF35KkMoqpMo',
+  authDomain: 'igbo-api-admin.firebaseapp.com',
+  databaseURL: 'https://igbo-api-admin.firebaseio.com',
+  projectId: 'igbo-api-admin',
+  storageBucket: 'igbo-api-admin.appspot.com',
+  messagingSenderId: '387052944949',
+  appId: '1:387052944949:web:1fc287146a86f425677b7c',
+  measurementId: 'G-HC6CF1YFN2',
+};
+
+export const STAGING_FIREBASE_CONFIG = {
+  apiKey: 'AIzaSyAD-Z6W84l75d5f1sEU9J4Uq0cXQtYS-Y0',
+  authDomain: 'igbo-api-admin-staging.firebaseapp.com',
+  databaseURL: 'https://igbo-api-admin-staging.firebaseio.com',
+  projectId: 'igbo-api-admin-staging',
+  storageBucket: 'igbo-api-admin-staging.appspot.com',
+  messagingSenderId: '456884948021',
+  appId: '1:456884948021:web:44069a1821ff4c8c34357c',
+  measurementId: 'G-TGHHWCYZJ6',
+};
 
 export default (): any => {
   const firebaseConfig = process.env.NODE_ENV === 'production' ? PROD_FIREBASE_CONFIG : STAGING_FIREBASE_CONFIG;
