+++ src/shared/constants/Collections.ts
@@ -6,7 +6,6 @@ enum Collections {
   WORD_SUGGESTIONS = 'wordSuggestions',
   EXAMPLE_SUGGESTIONS = 'exampleSuggestions',
   CORPUS_SUGGESTIONS = 'corpusSuggestions',
-  GENERIC_WORDS = 'genericWords',
   USERS = 'users',
   POLLS = 'polls',
   NOTIFICATIONS = 'notifications',
